restructure the whole transaction overview
