--- src/app/Features/csv/components/transaction-table/transaction-table.component.css
@@ -1,48 +1,0 @@
-table {
-    @apply
-    min-w-full select-none shadow-2xl rounded-lg border-[#f1f1f1] border font-outfit text-[#014e60];
-}
-
-thead {
-    @apply
-    uppercase bg-[#f1f1f1];
-}
-
-th {
-    @apply
-    px-4 py-2 text-left
-    select-text;
-}
-
-tbody{
-    @apply
-    text-gray-700 text-sm;
-}
-
-td {
-    @apply
-    border px-4 py-2
-    select-text;
-}
-
-input {
-    @apply
-    border p-1;
-}
-
-.edit {
-    @apply
-    font-medium
-    text-blue-600
-    hover:underline;
-}
-
-.save {
-    @apply
-    text-green-500 hover:underline mr-4;
-}
-
-.cancel {
-    @apply
-    text-red-500 hover:underline;
-}
