+++ src/app/Features/transaction-overview/components/transaction-table/components/table-row/components/autocomplete-dropdown/autocomplete-dropdown.component.css
