+++ src/app/Features/transaction-overview/components/transaction-table/components/table-row/components/autocomplete-dropdown/autocomplete-dropdown.component.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatInputModule } from '@angular/material/input';
-import { FilterCategoriesPipe } from '../../pipes/filter-categories.pipe';
+import { FilterCategoriesPipe } from '../../../../../../pipes/filter-categories.pipe';
 
 @Component({
   selector: 'app-autocomplete-dropdown',
