+++ src/app/Features/transaction-overview/components/transaction-table/components/table-row/table-row.component.css
