+++ src/app/Features/transaction-overview/pipes/filter-categories.pipe.ts
